feat(TBN-1370): extract nestjs provider explorer package
