+++ .bumpy/extract-nestjs-provider-explorer.md
@@ -1,0 +1,9 @@
+---
+"@wisemen/nestjs-provider-explorer": minor
+"@wisemen/nestjs-domain-events": patch
+"@wisemen/nestjs-typesense": patch
+"@wisemen/nestjs-nats": patch
+"@wisemen/pgboss-nestjs-job": patch
+---
+
+Added a shared NestJS provider explorer package and migrated existing packages to use it.
