+++ packages/api/nestjs-domain-events/package.json
@@ -28,6 +28,7 @@
     "release": "pnpx release-it"
   },
   "dependencies": {
+    "@wisemen/nestjs-provider-explorer": "workspace:^",
     "expect": "catalog:test",
     "sinon": "catalog:test"
   },
