+++ packages/api/nestjs-nats/lib/nats-application-factory.ts
@@ -9,8 +9,10 @@ import { getNatsSubscriberHandlerConfig, isNatsSubscriberHandler } from './subsc
 import { getNatsConsumerHandlerConfig, isNatsConsumerHandler } from './consumers/nats-consumer-handler.decorator.js'
 import { isNatsServiceEndpoint, getNatsServiceEndpointConfig } from './services/nats-service-endpoint.decorator.js'
 import { NATS_STREAMS_TOKEN } from './tokens.js'
-import type { NestjsProvider } from './providers/providers-explorer.js'
-import { ProvidersExplorer } from './providers/providers-explorer.js'
+import {
+  type NestjsProvider,
+  ProvidersExplorer
+} from '@wisemen/nestjs-provider-explorer'
 
 @Injectable()
 export class NatsApplicationFactory {
