+++ packages/api/nestjs-nats/lib/nats.module.ts
@@ -1,9 +1,9 @@
 import { type DynamicModule, Module, type OnApplicationBootstrap, type OnApplicationShutdown, type Type } from '@nestjs/common'
+import { ProvidersExplorerModule } from '@wisemen/nestjs-provider-explorer'
 import type { ClassConstructor } from 'class-transformer'
 import { NatsApplication } from './nats-application.js'
 import { NatsApplicationFactory } from './nats-application-factory.js'
 import { NATS_STREAMS_TOKEN } from './tokens.js'
-import { ProvidersExplorerModule } from './providers/providers-explorer.module.js'
 
 export interface NatsApplicationModuleOptions {
   modules: Type<unknown>[]
