+++ packages/api/nestjs-provider-explorer/eslint.config.js
@@ -1,0 +1,12 @@
+import eslintNestJSConfig from '@wisemen/eslint-config-nestjs'
+
+export default [
+  ...eslintNestJSConfig,
+  {
+    files: ['**/*.test.ts'],
+    rules: {
+      '@typescript-eslint/unbound-method': 'off',
+      '@typescript-eslint/no-floating-promises': 'off'
+    }
+  }
+]
