+++ packages/api/nestjs-provider-explorer/lib/index.ts
@@ -1,0 +1,3 @@
+export type { NestjsProvider } from './providers-explorer.js'
+export { ProvidersExplorer } from './providers-explorer.js'
+export { ProvidersExplorerModule } from './providers-explorer.module.js'
