+++ packages/api/nestjs-provider-explorer/lib/providers-explorer.module.ts
@@ -1,7 +1,9 @@
 import { Module } from '@nestjs/common'
+import { DiscoveryModule } from '@nestjs/core'
 import { ProvidersExplorer } from './providers-explorer.js'
 
 @Module({
+  imports: [DiscoveryModule],
   providers: [ProvidersExplorer],
   exports: [ProvidersExplorer]
 })
