+++ packages/api/nestjs-provider-explorer/tsconfig.json
@@ -1,0 +1,16 @@
+{
+  "extends": "../../../tsconfig.base.json",
+  "compilerOptions": {
+    "outDir": "./dist",
+    "rootDir": "./lib",
+    "types": [
+      "node"
+    ],
+    "paths": {
+      "#src/*": ["./lib/*"]
+    }
+  },
+  "include": [
+    "./lib/**/*.ts"
+  ]
+}
