--- packages/api/nestjs-typesense/lib/providers/providers-explorer.ts
@@ -1,45 +1,0 @@
-import { Injectable, type Type } from '@nestjs/common'
-import { ModulesContainer } from '@nestjs/core'
-
-export interface NestjsProvider {
-  providerClass: Type<unknown>
-  providerInstance: object
-}
-
-@Injectable()
-export class ProvidersExplorer {
-  private readonly discoveredProviders: NestjsProvider[] = []
-  private isInitialized = false
-
-  constructor (private readonly modules: ModulesContainer) {}
-
-  get providers (): NestjsProvider[] {
-    if (!this.isInitialized) {
-      this.initialize()
-      this.isInitialized = true
-    }
-
-    return Array.from(this.discoveredProviders)
-  }
-
-  private initialize (): void {
-    for (const moduleWrapper of this.modules.values()) {
-      for (const providerWrapper of moduleWrapper.providers.values()) {
-        const providerClass = providerWrapper.metatype
-
-        if (providerClass == null) {
-          continue
-        }
-
-        if (!Object.hasOwn(providerClass, 'prototype')) {
-          continue
-        }
-
-        this.discoveredProviders.push({
-          providerClass: providerClass as Type<unknown>,
-          providerInstance: providerWrapper.instance as object
-        })
-      }
-    }
-  }
-}
