+++ packages/api/nestjs-typesense/package.json
@@ -34,7 +34,9 @@
     "@wisemen/pagination": "workspace:^",
     "@wisemen/coordinates": "workspace:^"
   },
-  "dependencies": {},
+  "dependencies": {
+    "@wisemen/nestjs-provider-explorer": "workspace:^"
+  },
   "devDependencies": {
     "@types/node": "catalog:backend",
     "@wisemen/eslint-config-nestjs": "workspace:*",
