--- packages/api/pgboss-nestjs-job/lib/providers/providers-explorer.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common'
-import { ProvidersExplorer } from './providers-explorer.js'
-
-@Module({
-  providers: [ProvidersExplorer],
-  exports: [ProvidersExplorer]
-})
-export class ProvidersExplorerModule {}
