+++ packages/api/pgboss-nestjs-job/lib/worker/pgboss-worker.module.ts
@@ -1,8 +1,8 @@
 import assert from 'assert'
 import { DynamicModule, Module } from '@nestjs/common'
+import { ProvidersExplorerModule } from '@wisemen/nestjs-provider-explorer'
 import { JobModule } from '../jobs/job.module.js'
 import { PgBossClientModule } from '../client/pgboss-client.module.js'
-import { ProvidersExplorerModule } from '../providers/providers-explorer.module.js'
 import { ConfigurableModuleClass, PgbossWorkerModuleAsyncOptions, PgbossWorkerModuleOptions } from './pgboss-worker.module-definition.js'
 import { PgbossBouncerRegistry } from './pgboss-bouncer.registry.js'
 import { PgbossWorkerApp } from './pgboss-worker-app.js'
